Fix error handling in OS.Main.run
We printed the exception but returned unit, causing the unikernel to
exit cleanly. This didn't trigger the on_crash behaviour of preserving
the domain, meaning that the console output would be lost.

Instead, simply let the exception propagate back to the C code, which
already handles exceptions.

Note: on x86_64, Mini-OS tries to print a stack trace in do_exit for
some reason, which results in a "Page fault" message being printed
too, but that can be fixed separately.

See: https://github.com/mirage/xen-arm-builder/issues/46