Drop support for old ocaml-gnt package
This allows us to remove some unused C stubs and means that we are now
free to update the C functions without having to keep them in sync with
ocaml-gnt.

The list of free grant refs is now kept in this module (the new code is
just copied from ocaml-gnt, with the tests for Unix removed).

Attempting to link a unikernel that uses ocaml-gnt will now fail with
e.g.

    undefined reference to `stub_gnttab_init'

This is intended, since we would otherwise have two modules trying to
track the free grant IDs. ocaml-gnt has marked all of the Xen-specific
functions as deprecated, so users should get a more useful error from
the compiler before getting to this stage, unless they used only the
part of the API shared with Unix.