Hook up OS.MM.Heap_pages.{total,used}
The total number of heap pages is computed from the Solo5 heap size
(si->heap_size) and the number of used pages is computed from the
result returned by dlmalloc's malloc_footprint().

Note that mallinfo() is deliberately not used in the latter case as that
call is expensive and existing unikernels rely on this interface being
cheap to call.