OS.MM: Replace with new OS.Memory interface
Replace the OS.MM module with a new OS.Memory statistics interface
modeled on the stdlib's Gc stat type. All units are now in the system
word size rather than pages which are arch-specific and ill-defined.

The implementation takes care to account for the fact that on Solo5
backends the heap is shared with the stack, and a malloc() allocation on
the heap will fail if the heap would grow to within 1MB of the stack.