Remove OS.Xen.virt_to_mfn
This function is a vestige from the Mini-OS days, and in the new
Solo5-based Xen stack is effectively (addr >> 12) since all virtual
addresses are identity mapped.

Applications wanting to do this kind of low-level manipulation should do
the computation themselves. We may consider re-adding a more thought out
interface in the future, if necessary.