add caml_get_wall_clock -- similar to caml_get_monotonic_time
The motivation is to avoid lots of conversions in mirage-clock just to retrieve
the current wall clock time. This is one part to solve
https://github.com/mirage/mirage-clock/issues/23

The unix_gettimeofday is still useful for the OCaml runtime system - which uses
it to initialize its RNG (the alternative is it being initialized with 0s which
is obviously worse) - that's why that symbol is kept as well.